+++ forui_hooks/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.17.0
 * **Breaking** Remove `useFSelectTileGroupController`. Use `useFSelectGroupController` instead.
+* **Breaking** Remove `useFSelectMenuTileController`. Use `useFSelectGroupController` instead.
+* **Breaking** Remove `useFRadioSelectMenuTileController`. Use `useFRadioSelectGroupController` instead.
 
 
 ## 0.16.0